+++ __tests__/components/atoms/CircleProgressBar.test.js
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import { axe, toHaveNoViolations } from 'jest-axe'
-import CircleProgressBar from './CircleProgressBar'
+import CircleProgressBar from '../../../components/atoms/CircleProgressBar'
 
 expect.extend(toHaveNoViolations)
